Move networks to NetworkDefinitions
