+++ src/Stratis.Bitcoin.Features.SmartContracts/Networks/SmartContractsRegTest.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using NBitcoin;
 using NBitcoin.DataEncoders;
-using NBitcoin.Networks;
+using NBitcoin.NetworkDefinitions;
 using NBitcoin.Protocol;
 
 namespace Stratis.Bitcoin.Features.SmartContracts.Networks
@@ -15,8 +15,8 @@ namespace Stratis.Bitcoin.Features.SmartContracts.Networks
         public SmartContractsRegTest()
         {
             this.Name = "SmartContractsRegTest";
-            this.RootFolderName = NBitcoin.Networks.StratisMain.StratisRootFolderName;
-            this.DefaultConfigFilename = NBitcoin.Networks.StratisMain.StratisDefaultConfigFilename;
+            this.RootFolderName = NBitcoin.NetworkDefinitions.StratisMain.StratisRootFolderName;
+            this.DefaultConfigFilename = NBitcoin.NetworkDefinitions.StratisMain.StratisDefaultConfigFilename;
             this.Magic = 0xDAB5BFFA;
             this.DefaultPort = 18444;
             this.RPCPort = 18332;
