+++ src/Stratis.Bitcoin.Features.SmartContracts/Networks/SmartContractsTest.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using NBitcoin;
 using NBitcoin.DataEncoders;
-using NBitcoin.Networks;
+using NBitcoin.NetworkDefinitions;
 using NBitcoin.Protocol;
 
 namespace Stratis.Bitcoin.Features.SmartContracts.Networks
@@ -12,8 +12,8 @@ namespace Stratis.Bitcoin.Features.SmartContracts.Networks
         public SmartContractsTest()
         {
             this.Name = "SmartContractsTestNet";
-            this.RootFolderName = NBitcoin.Networks.StratisMain.StratisRootFolderName;
-            this.DefaultConfigFilename = NBitcoin.Networks.StratisMain.StratisDefaultConfigFilename;
+            this.RootFolderName = NBitcoin.NetworkDefinitions.StratisMain.StratisRootFolderName;
+            this.DefaultConfigFilename = NBitcoin.NetworkDefinitions.StratisMain.StratisDefaultConfigFilename;
             this.Magic = 0x0709110E; // Incremented 19/06
             this.DefaultPort = 18333;
             this.RPCPort = 18332;
